+++ app/src/main/java/com/print/color/printcolor/ui/quotationList/QuotationListItem.kt
@@ -1,0 +1,54 @@
+package com.print.color.printcolor.ui.quotationList
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.RowScope
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import com.print.color.printcolor.R
+import com.print.color.printcolor.domain.model.Quotation
+
+@Composable
+fun RowScope.QuotationListItem(quotation: Quotation) {
+    Card(
+        modifier = Modifier
+            .padding(all = 8.dp)
+            .border(
+                BorderStroke(1.dp, MaterialTheme.colorScheme.onBackground),
+                shape = RoundedCornerShape(8.dp)
+            )
+            .weight(1f),
+        colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surfaceVariant)
+    ) {
+        Row(modifier = Modifier
+            .fillMaxWidth()
+            .padding(all = 16.dp)) {
+            Column(
+                modifier = Modifier.weight(1f)
+            ) {
+                Text("ID: ${quotation.id}", fontWeight = FontWeight.Bold)
+                Text("Contact: ${quotation.contact}")
+                Text("Client: ${quotation.clientName}")
+            }
+            PcSStepImage(
+                icon = R.drawable.ic_pcs_survey,
+                contentDescription = "",
+                modifier = Modifier.align(
+                    Alignment.CenterVertically
+                )
+            )
+        }
+    }
+}
