+++ app/src/main/java/com/print/color/printcolor/ui/quotationList/QuotationListScreen.kt
@@ -1,7 +1,6 @@
 package com.print.color.printcolor.ui.quotationList
 
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.border
+import android.util.Log
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -9,21 +8,13 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.rounded.Search
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
@@ -32,10 +23,8 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
@@ -44,9 +33,9 @@ import com.print.color.printcolor.domain.model.Quotation
 import com.print.color.printcolor.ui.components.TextFieldTheme.PcsTextField
 import com.print.color.printcolor.ui.components.TextFieldTheme.model.TextFieldData
 import com.print.color.printcolor.ui.components.TextFieldTheme.model.TextFieldType.OUTLINED
-import com.print.color.printcolor.ui.components.ToolTip.PcSRichTooltip
-import com.print.color.printcolor.ui.components.ToolTip.model.ToolTipData
 import com.print.color.printcolor.ui.theme.PrintColorTheme
+import kotlin.collections.chunked
+import kotlin.collections.forEach
 
 
 /*@Preview(
@@ -81,19 +70,11 @@ fun QuotationListScreen(
                         keyboardType = KeyboardType.Text,
                         leadingIcon = painterResource(R.drawable.ic_pcs_search)
                     ),
-                    modifier = Modifier.weight(1f),
+                    modifier = Modifier,
                     onValueChange = { searchBarText = it },
                     value = searchBarText,
                     imeAction = ImeAction.Search
                 )
-                IconButton(onClick = {}) {
-                    Icon(
-                        modifier = modifier,
-                        painter = painterResource(R.drawable.ic_pcs_mode_list),
-                        contentDescription = "view mode list",
-                        tint = MaterialTheme.colorScheme.onTertiary
-                    )
-                }
                 IconButton(onClick = {}) {
                     Icon(
                         modifier = modifier,
@@ -108,140 +89,48 @@ fun QuotationListScreen(
                     .fillMaxWidth()
                     .padding(vertical = 16.dp)
             )
-            PcSSteps()
-            QuotationList(uiState.isLoading, uiState.quotations, searchBarText)
+            QuotationListSteps() // Quotation List Steps
+            /*val filteredData = uiState.quotations.filter { it.id.contains(searchBarText, ignoreCase = true) }
+            if (uiState.isLoading) {
+                Log.d("QuotationList", "Loading...")
+                CircularProgressIndicator()
+            } else {
+                Log.d("QuotationList", "Loaded")
+                LazyColumn(
+                    modifier = Modifier
+                        .padding(16.dp)
+                        .fillMaxSize()
+                ) {
+                    items(filteredData.chunked(2)) { rowItems ->
+                        Row(
+                            modifier = Modifier.fillMaxWidth(),
+                            horizontalArrangement = Arrangement.spacedBy(8.dp)
+                        ) {
+                            rowItems.forEach { quotation ->
+                                QuotationListItem(quotation = quotation)
+                            }
+                            if (rowItems.size < 2) {
+                                Spacer(modifier = Modifier.weight(1f))
+                            }
+                        }
+                    }
+                }
+            }*/
+            QuotationList(uiState.isLoading, uiState.quotations, searchBarText) // Quotation List
         }
     }
 }
 
-//@Preview(showBackground = true)
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun PcSSteps(modifier: Modifier = Modifier) {
-    PrintColorTheme {
-        Card(
-            colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surfaceVariant),
-            modifier = modifier
-                .border(
-                    BorderStroke(1.dp, MaterialTheme.colorScheme.onBackground),
-                    shape = RoundedCornerShape(8.dp)
-                )
-                .fillMaxWidth()
-        ) {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(all = 16.dp),
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.SpaceEvenly
-            ) {
-                Column(
-                    modifier = Modifier,
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    PcSStepImage(icon = R.drawable.ic_pcs_notes, contentDescription = "")
-                    PcSRichTooltip(
-                        modifier = modifier,
-                        data = ToolTipData(
-                            toolTipTitle = stringResource(R.string.quotation_list_screen_tooltip_title_step1),
-                            toolTipDescription = stringResource(R.string.quotation_list_screen_tooltip_description_step1),
-                            toolTipActionText = stringResource(R.string.quotation_list_screen_tooltip_action_step1)
-                        )
-                    )
-                }
-                PcsStepDivider()
-                Column(
-                    modifier = Modifier,
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    PcSStepImage(icon = R.drawable.ic_pcs_success, contentDescription = "")
-                    PcSRichTooltip(
-                        data = ToolTipData(
-                            toolTipTitle = stringResource(R.string.quotation_list_screen_tooltip_title_step2),
-                            toolTipDescription = stringResource(R.string.quotation_list_screen_tooltip_description_step2),
-                            toolTipActionText = stringResource(R.string.quotation_list_screen_tooltip_action_step2)
-                        )
-                    )
-                }
-                PcsStepDivider()
-                Column(
-                    modifier = Modifier,
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    PcSStepImage(icon = R.drawable.ic_pcs_design, contentDescription = "")
-                    PcSRichTooltip(
-                        data = ToolTipData(
-                            toolTipTitle = stringResource(R.string.quotation_list_screen_tooltip_title_step3),
-                            toolTipDescription = stringResource(R.string.quotation_list_screen_tooltip_description_step3),
-                            toolTipActionText = stringResource(R.string.quotation_list_screen_tooltip_action_step3)
-                        )
-                    )
-                }
-                PcsStepDivider()
-                Column(
-                    modifier = Modifier,
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    PcSStepImage(icon = R.drawable.ic_pcs_print, contentDescription = "")
-                    PcSRichTooltip(
-                        data = ToolTipData(
-                            toolTipTitle = stringResource(R.string.quotation_list_screen_tooltip_title_step4),
-                            toolTipDescription = stringResource(R.string.quotation_list_screen_tooltip_description_step4),
-                            toolTipActionText = stringResource(R.string.quotation_list_screen_tooltip_action_step4)
-                        )
-                    )
-                }
-                PcsStepDivider()
-                Column(
-                    modifier = Modifier,
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    PcSStepImage(icon = R.drawable.ic_pcs_delivery, contentDescription = "")
-                    PcSRichTooltip(
-                        data = ToolTipData(
-                            toolTipTitle = stringResource(R.string.quotation_list_screen_tooltip_title_step5),
-                            toolTipDescription = stringResource(R.string.quotation_list_screen_tooltip_description_step5),
-                            toolTipActionText = stringResource(R.string.quotation_list_screen_tooltip_action_step5)
-                        )
-                    )
-                }
-                PcsStepDivider()
-                Column(
-                    modifier = Modifier,
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    PcSStepImage(icon = R.drawable.ic_pcs_survey, contentDescription = "")
-                    PcSRichTooltip(
-                        data = ToolTipData(
-                            toolTipTitle = stringResource(R.string.quotation_list_screen_tooltip_title_step6),
-                            toolTipDescription = stringResource(R.string.quotation_list_screen_tooltip_description_step6),
-                            toolTipActionText = stringResource(R.string.quotation_list_screen_tooltip_action_step6)
-                        )
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun PcsStepDivider(modifier: Modifier = Modifier) {
-    HorizontalDivider(
-        modifier = modifier
-            .padding(horizontal = 8.dp)
-            .size(height = 2.dp, width = 50.dp),
-        color = Color.Black
-    )
-}
-
-
+/** Quotation List */
 @Composable
 fun QuotationList(isLoading: Boolean, quotations: List<Quotation>, searchBarText: String) {
     val filteredData = quotations.filter { it.id.contains(searchBarText, ignoreCase = true) }
 
     if (isLoading) {
+        Log.d("QuotationList", "Loading...")
         CircularProgressIndicator()
     } else {
+        Log.d("QuotationList", "Loaded")
         LazyColumn(
             modifier = Modifier
                 .padding(16.dp)
@@ -253,35 +142,7 @@ fun QuotationList(isLoading: Boolean, quotations: List<Quotation>, searchBarText
                     horizontalArrangement = Arrangement.spacedBy(8.dp)
                 ) {
                     rowItems.forEach { quotation ->
-                        Card(
-                            modifier = Modifier
-                                .padding(all = 8.dp)
-                                .border(
-                                    BorderStroke(1.dp, MaterialTheme.colorScheme.onBackground),
-                                    shape = RoundedCornerShape(8.dp)
-                                )
-                                .weight(1f),
-                            colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surfaceVariant)
-                        ) {
-                            Row(modifier = Modifier
-                                .fillMaxWidth()
-                                .padding(all = 16.dp)) {
-                                Column(
-                                    modifier = Modifier.weight(1f)
-                                ) {
-                                    Text("ID: ${quotation.id}", fontWeight = FontWeight.Bold)
-                                    Text("Contact: ${quotation.contact}")
-                                    Text("Client: ${quotation.clientName}")
-                                }
-                                PcSStepImage(
-                                    icon = R.drawable.ic_pcs_survey,
-                                    contentDescription = "",
-                                    modifier = Modifier.align(
-                                        Alignment.CenterVertically
-                                    )
-                                )
-                            }
-                        }
+                        QuotationListItem(quotation = quotation)
                     }
                     if (rowItems.size < 2) {
                         Spacer(modifier = Modifier.weight(1f))
